Airflow/requirements.txt
apache-airflow
numpy
scikit-learn
pandas
pickle